update README
* try to fix ci

+++ src/oop/oop/entities/GameObject.h
@@ -8,7 +8,7 @@ namespace ecs::benchmarks::oop::entities {
 
     class GameObject {
     public:
-        using game_object_id_t = int64_t;
+        using game_object_id_t = size_t;
 
         GameObject() = default;
         GameObject(const GameObject &) = default;
